requirements.txt
pandas
torch
transformers
sklearn
absl-py
tensorflow
keras
sh
numpy
pytorch_lightning
